+++ app/src/androidTest/java/com/example/superapp/test/AlternativeSetupTest.kt
@@ -8,13 +8,7 @@ import androidx.test.platform.app.InstrumentationRegistry.getInstrumentation
 import com.dropbox.dropshots.Dropshots
 import com.dropbox.dropshots.ThresholdValidator
 import com.example.superapp.utils.CustomComparator
-import com.example.superapp.utils.awaitUntilShimmerDisappears
-import com.example.superapp.utils.awaitUntilWebviewAppears
-import com.example.superapp.utils.clickButtonWith
-import com.example.superapp.utils.delayFor
-import com.example.superapp.utils.goBack
 import com.example.superapp.utils.paywallPresentsFor
-import com.example.superapp.utils.screenshotFlow
 import com.superwall.sdk.Superwall
 import com.superwall.sdk.config.options.SuperwallOptions
 import com.superwall.superapp.Keys
@@ -23,7 +17,6 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
-import kotlin.time.Duration.Companion.milliseconds
 
 @RunWith(AndroidJUnit4::class)
 class AlternativeSetupTest {
@@ -46,6 +39,9 @@ class AlternativeSetupTest {
         }
     }
 
+    /*
+    Flaky test due to using old paywall
+
     @Test
     fun test_paywall_displays_on_session_start() =
         with(dropshots) {
@@ -59,6 +55,7 @@ class AlternativeSetupTest {
             )
             paywallPresentsFor(UITestHandler.test50Info)
         }
+     */
 
     @Test
     fun test_paywall_displays_on_app_install() =
@@ -87,7 +84,10 @@ class AlternativeSetupTest {
             )
             paywallPresentsFor(UITestHandler.test53Info)
         }
-/*
+    /*
+    Flaky test due to using old paywall
+
+
     @Test
     fun test_paywall_displays_on_deep_link() =
         with(dropshots) {
@@ -102,7 +102,6 @@ class AlternativeSetupTest {
             paywallPresentsFor(UITestHandler.test57Info)
         }
 
- */
 
     @Test
     fun test_paywall_displays_on_decline() =
@@ -133,4 +132,6 @@ class AlternativeSetupTest {
                 }
             }
         }
+
+     */
 }