+++ superwall/src/androidTest/java/com/superwall/sdk/store/transactions/TransactionManagerTest.kt
@@ -182,6 +182,7 @@ class TransactionManagerTest {
             storage = storage,
             entitlementsById = entitlementsById,
             showRestoreDialogForWeb = showRestoreDialogForWeb,
+            refreshReceipt = {},
         )
     }