+++ superwall/src/main/java/com/superwall/sdk/store/AutomaticPurchaseController.kt
@@ -166,8 +166,8 @@ class AutomaticPurchaseController(
                 .also {
                     // Do not set empty offer token for one time products
                     // as Google play is not supporting it since June 12th 2024
-                    if (!isOneTime) {
-                        it.setOfferToken(offerToken ?: "")
+                    if (!isOneTime && offerToken != null) {
+                        it.setOfferToken(offerToken)
                     }
                 }.build()
 
@@ -199,9 +199,7 @@ class AutomaticPurchaseController(
             BillingFlowParams
                 .newBuilder()
                 .apply {
-                    if (shouldPassIdToPlayStore && id != null) {
-                        setObfuscatedAccountId(id)
-                    }
+                    setObfuscatedAccountId(Superwall.instance.externalAccountId)
                 }.setProductDetailsParamsList(listOf(productDetailsParams))
                 .build()